2038: separate the ECAA functionality from the ECA (eca.go -> ecaa.go)
